Upload EXPLAIN output as a snippet
EXPLAIN plans can quickly get rather large. In such cases chatops will
respond with a link to the build as its too large to send back to Slack.
To work around this we simply upload the output of EXPLAIN as a snippet.
This has the added benefit of taking up less space unless a user
explicitly opens the snippet.