Don't fetch repo each chatops command
Prior to this change, each chatops command would use git to sync up its
state. This is a wasteful process as between commands nothing will
change.

This change sets the strategy to `none` so the runner will skip that.
The docker images now must have the repository baked in, which is done
at the build stage.