Copy project file to relative location to WORKDIR
Starting the script didn't work as we copied to a relative path, so the
paths shifted when executing.

A few CI job variables are shifted too, this makes the CI spec more
portable and since 9.0 the X_BUILD_Y vars where duplicated to X_JOB_Y.
Where some day the former will be removed, this will make sure we never
have to worry about it.