Add `Release::Command.trigger_release`
This is a helper method that will perform the Run Trigger > Fetch
Pipelines > Check for chatops > Print result workflow common to all
Release commands.

If given a block it will return a TriggerResult object with the status
and the resulting URL (either a job, or a pipeline). Otherwise it will
print the default message depending on status.