Building images on the ops instance
This adds support for building the Docker image on the ops instance.
This works by pushing into whatever the local registry is, instead of
depending on a hardcoded registry URL.

For running chatops jobs we still use a hardcoded value, as we always
want to use the ops instance's image.