Add the prepare command
This triggers the release-tools `patch_merge_request` command, which is
named a little verbosely, so we just call this command `prepare`, since
it's creating preparation MRs.

In the future, this could be an all-in-one command to create the task
list issue and then the prep MRs, assuming those are always tied
together.