Support `--master` flag in `release:merge`
Since this flag is only useful for the single command, it didn't make
sense to have it as a top-level command option, which unfortunately
forces us to handle our own command parsing rather than relying on Slop.

This also updates the `run_trigger` helper to accept an optional
`params` hash, which will get extended with the default values we send
on every release trigger.