Fix merging security MRs targeting master
The release tools project doesn't care about the value of the variable,
and only checks if it is empty or not. Passing booleans directly would
result in "false" being passed when omitting the --master flag. This
wouuld then result in the release tools project merging MRs targeting
master, because "false" evaluates to true.

To fix this, we pass an empty string when not using the --master flag.