Log changing the value and scopes of feature flags
Whenever the value and/or scope of a feature flag is changed, we create
a closed issue in a dedicated project. This issue contains some
information about who changed the feature flag, what it was changed to,
when this took place, what host was affected, etc. Using these issues we
no longer need to dig through tons of messages in Slack to figure out
who changed what.

Fixes https://gitlab.com/gitlab-com/gl-infra/delivery/issues/525