Prevent changing feature flags during incidents
This changes the "feature set" command so it checks for any ongoing
incidents in the issue tracker
https://gitlab.com/gitlab-com/gl-infra/production/-/issues. If there are
one or more S1, S2, or S3 issues, the state of a feature flag can't be
changed.

In the rare case that somebody needs to work around this, they can
specify the --ignore-incidents option. When this option is set we add
some additional data to the feature flag log, making it clear the user
chose to ignore the incidents.

This fixes https://gitlab.com/gitlab-com/chatops/-/issues/81