Updates canary to use server name instead of chef role
For the canary chatops command we previously were looking up servers
based on all nodes that were in the chef role for canary omnibus.

This change updates that logic so instead we identify canary servers
as any server with '-cny-' in the name, this will also cover GKE
which lives outside of Chef.