Fix usage of sections in auto-deploy blockers
When I first tested this code I used the yield API and it worked. But
when writing tests I changed this to a return based API, believing
(based on Slack's source code) this would work. It does not, as
`blocks.context` returns the block; not the context.

This commit fixes this by changing the code to use the yield API, and
the tests to test for this correctly.