Remove stable branch release command
As part of the epic
https://gitlab.com/groups/gitlab-com/gl-infra/-/epics/236, we added code
to Release Tools that can perform releases using the API, including
creating stable branches. In merge request
https://gitlab.com/gitlab-org/release-tools/-/merge_requests/1217 we
remove the Release Tools code for creating stable branches, as this is
no longer necessary.

In this commit, we also remove the "stable_branch" command and related
code, as this command too is no longer necessary.