Treat unmatched args as an error in deploy command
This shifts the deploy version off of the arguments array, and if any
additional (unprocessed) options are remaining, we treat it as an error
state, returning early.

This prevents the case where a user provides an unquoted argument to
`--ignore-production-checks`, resulting in a message that looks to be
cut off.

For example, with this command:

  /chatops run deploy 13.5.0 --ignore-production-checks Some reason to skip

the previous implementation would have only considered "Some" as the
reason, ignoring the rest of the input.