Only consider a deployment as running if it's first
If our deployment tracking fails for any reason, a deployment might be
left in a "running" state in GitLab. When the `rollback check` command
goes to see if there are any running deployments, it filters by the
`running` status, which will return a deployment with that status
regardless of when it was updated, leading to misleading information.

Now, we fetch the first deployment regardless of status, and if that one
happens to be running, we return it. Otherwise no running deployment is
returned.