Always set IGNORE_PRODUCTION_CHECKS on a rollback
During https://gitlab.com/gitlab-com/gl-infra/production/-/issues/3852
execution we realized that we will always run a rollback in a
situation where at least a deployment blocker is present.

There will be an active incident and the omnibus chef role could be
locked.

When we send a chatops rollback command, it should set
`IGNORE_PRODUCTION_CHECKS` variable to `This is a rollback` to skip
the checks and leave the proper note in the monthly issue.

Related to: https://gitlab.com/gitlab-com/gl-infra/delivery/-/issues/1735