Add a command to check if MR will be published in SM release
The command checks if the MR is present in the stable branch. If the
MR is present in the stable branch, the MR will be published in the
SM release.

If the stable branch doesn't exist yet, it will check if the MR
has been deployed to production. If the MR has been deployed to
gprd, it will most likely be included in the stable branch when it
is created.