Check tag before checking stable branch
If there already is a tag for the given release version, the release
has already happened, and we can use past tense in the response
message. For example, "MR was released in v14.2.0-ee".

Make the version an optional argument to the command since we only
need it for the stable branch check. If the MR exists in a tag, we
won't check the stable branch, so the version won't be required.