Ensure consistency between staging and production environments
This change checks the status of the feature flag on the
other envrionment (production/staging). When turning the FF on
in production, ideally the feature flag should be turned on in
staging to test it. When turning off in staging, the feature flag
should have off in production too.

Issue: gitlab.com/gitlab-com/chatops/-/issues/109
MR: gitlab.com/gitlab-com/chatops/-/merge_requests/241