Swap the comparison link for an upcoming deployment
When a deployment is running, the previous behavior generates an empty
compare link, because we're comparing a new version to a previous
version.

Now if the "current" version is actually upcoming (i.e., running), we
need to swap the order so that we compare the previous version
to the new version.