Deploy command takes an explicit environment argument
Previously we'd operate on one or more flags to set the environment
(e.g., `--canary --production` deploys to `gprd-cny`). With the
introduction of `staging-ref` and `gstg-cny` this would get unwieldy.

Now the deploy command requires an explicit, familiar-looking
environment argument, such as `gprd`, `gstg-cny`, etc.

For user friendliness, we support the environment being before or after
the version argument.