Add semantic_logger to chatops
Log every GitLab API request. This should make it easier to see what
requests are being made, and to debug failures.

This also makes it possible for developers to log any relevant
information. The logs will appear in the job logs, and we can also
configure them to be uploaded to Elastic.