Fix the auto_deploy status command to work with cherry-picked commits
Currently, the command checks if an auto deploy branch has been
deployed. But with commits cherry-picked into an auto deploy branch,
it is possible for one or more of the latest commits in the auto
deploy branch to not be included in the deployment.

This commit updates the command to also check if the SHA being checked
is chronologically before the deployed SHA in the auto deploy branch.
If it is before the deployed SHA, then it was deployed to the
environment.