Check status of job in check_production pipeline
After triggering a pipeline in release tools to check the status
of production, look at the job status instead of the pipeline
status to know if the check failed or succeeded.

On the release-tools pipeline config, the job will be changed to be
allowed to fail. This means that the pipeline will succeed, but the
job will fail.

This helps prevent unnecessary Slack notifications about failing
pipelines.