Chatops commands for rolling out Zoekt
In https://gitlab.com/gitlab-org/gitlab/-/merge_requests/116650 we added
some admin APIs that will be used during our production rollout of
Zoekt. This MR makes just 2 of these endpoints accessible in chatops as
these are the ones that require write access. The other data can be
easily accessed from read-only consoles if needed. We may consider
exposing those via chatops eventually but for now these are the ones we
need to move the rollout forward.