Fix feature definition search matching when there are patch files
As per
https://docs.gitlab.com/ee/development/feature_flags/#optionally-add-a-patch-file-for-automated-removal-of-feature-flags
we allow developers to add a patch file accompanying their feature
flags. This patch file goes in the same directory as the YML definition
with the same name but the `.patch` extension. This causes problems with
the regex search added in
https://gitlab.com/gitlab-com/chatops/-/merge_requests/507 because it
matches any file with this name in these directories but does not
consider the extension.

This was causing my FF enables to fail in
https://ops.gitlab.net/gitlab-com/chatops/-/jobs/14037907 for the search
https://gitlab.com/search?search=allow_ai_features_token_for_graphql_ai_features+file%3A%5E%28config%2Ffeature_flags%2F%7Cee%2Fconfig%2Ffeature_flags%2F%29&nav_source=navbar&project_id=278964&group_id=9970&search_code=true&repository_ref=master
.

This MR fixes that by changing the search to include the extension and
the resulting search is
https://gitlab.com/search?search=%2Fallow_ai_features_token_for_graphql_ai_features.yml%24+file%3A%5E%28config%2Ffeature_flags%2F%7Cee%2Fconfig%2Ffeature_flags%2F%29&nav_source=navbar&project_id=278964&group_id=9970&search_code=true&repository_ref=master
which matches exactly my YAML file.

The other thing I changed was to add `/` to the start of the search.
This will mean that we don't get partially matching file names. Because
Zoekt will just do substring matching by default so if you had a feature
flag called `allow_users_to_do_stuff` and another one called `do_stuff`
then the search for the `do_stuff` feature flag would find both of them.
With the leading `/` we now confirm it matches the full file name.